==> Code/6_within_state_dependencies.Rmd <==
---
title: 'within state dependencies'
author: "Laura Symul"
date: "8/11/2020"
output: html_document
---


# Learning within-state dependencies

In this section, we show how our model is able to learn within-state dependencies between variables and to accurately decode the sequence of hidden states when the only difference between the two states is the direction of the correlation between the variables.

To show this, we proceed as follows.
We first specify a two-state HSMM decoding time-series of two continuous variables whose marginal emission probabilities are identical in the two states. At specification, these two variables are independent. We then simulate a time-series from this specified model. We call this time-series `X1`. The next step consists in introducing within-state correlation between the variables in this time-series. The modified time-series is named `X2`. We can now train the specified model on this modified time-series. Finally, we decode the modified time-series with the fitted model and compare the decoding accuracy with the decoding with the specified (not fitted) model.


```{r within-state-dep-setup, include = FALSE}

library(HiddenSemiMarkov)
library(tidyverse)

```

__Model specification__

```{r within-state-dep-spec}

m = specify_hsmm(J = 2,
                 init = c(1,0),
                 transition = matrix(c(0,1,1,0), nrow = 2, ncol = 2),
                 sojourn = list(type = "gamma", shape = c(10,10), scale = c(1,1)),
                 marg_em_probs = list(
                   var1 = list(type = "norm", params = list(mean = c(0,0), sd = c(1,1))),
                   var2 = list(type = "norm", params = list(mean = c(0,0), sd = c(5,5)))
                 ),
                 state_names = c("A","B"),
                 state_colors = c("lightskyblue2", "lightskyblue4"))
```


```{r within-state-dep-sim-viz, fig.width=8, fig.height= 3, fig.cap = "Simulated data with the specified model."}

X1 = simulate_hsmm(model = m, n_state_transitions = 20)
plot_hsmm_seq(X1, model = m, add_state_color_legend = TRUE)

```


__Modifying time-series to introduce within-state dependencies__

```{r within-state-dep-modifying-data}

X2 = X1
var1_state1 = sort(X2$var1[X2$state == 1])
var2_state1 = sort(X2$var2[X2$state == 1])
j = sample(1:length(var1_state1))
var1_state1 = var1_state1[j]
var2_state1 = var2_state1[j]

var1_state2 = sort(X2$var1[X2$state == 2])
var2_state2 = sort(X2$var2[X2$state == 2], decreasing = TRUE)
j = sample(1:length(var1_state2))
var1_state2 = var1_state2[j]
var2_state2 = var2_state2[j]

X2$var1[X2$state == 1] = var1_state1
X2$var2[X2$state == 1] = var2_state1
X2$var1[X2$state == 2] = var1_state2
X2$var2[X2$state == 2] = var2_state2

```


```{r within-state-dep-mod-seq-dep, fig.width=8, fig.height= 3, fig.cap = "Modified data to introduce within-state correlations."}

plot_hsmm_seq(X2, model = m, add_state_color_legend = TRUE)

```


```{r within-state-dep-mod-data-viz, fig.width=8, fig.height= 4, fig.cap = "Scatter-plot between the two variables of the original and modified time-series."}

Xb = bind_rows(
  X1 %>%  mutate(data_type = "simulated data"),
  X2 %>%  mutate(data_type = "modified data")
) %>% 
  mutate(state = factor(state),
         data_type = data_type %>% factor(., levels = c("simulated data","modified data")))


ggplot(Xb, aes(x = var1, y = var2, col = state))+
  geom_point() + scale_color_manual(values = m$state_colors) +
  facet_grid( . ~ data_type)
  
```



__Fitting the model to the modified time-series__

```{r within-state-dep-fit, fig.width=3, fig.height=3, fig.cap="Fitting the specified model to the modified data: likelihood at each EM iteration."}

model_fit = fit_hsmm(model = m, X = X2, rel_tol = 1/10000)
plot_hsmm_fit_status(fit_output = model_fit)

```

__Predicting sequence of hidden states with both models: the specified one and the fitted one__

```{r within-state-dep-predict}

vit_original = predict_states_hsmm(model = m, X = X2, method = "Viterbi")

vit_fit = predict_states_hsmm(model = model_fit$model, X = X2, method = "Viterbi")

```

__Performances__

Decoding accuracy on the modified data with the specified model: `r mean(vit_original$state_seq$state == X1$state)`

Decoding accuracy on the modified data with the fitted model: `r mean(vit_fit$state_seq$state == X1$state)`

```{r within-state-dep-join-em, fig.height=6, fig.width=6, out.width='60%', fig.cap="Joint emission probability densities per state and model (normalized densities)."}

b_original = m$b %>% filter(!is.na(var1), !is.na(var2)) %>% 
  pivot_longer(col = starts_with("p_"), 
               names_to = "state", values_to = "prob", 
               names_prefix = "p_")

b_fit = model_fit$model$b %>% filter(!is.na(var1), !is.na(var2))%>% 
  pivot_longer(col = starts_with("p_"), 
               names_to = "state", values_to = "prob", 
               names_prefix = "p_")

b = bind_rows(
  b_original %>% mutate(model = "specified") %>% group_by(state) %>% 
    mutate(rel_prob = prob/max(prob)),
  b_fit %>% mutate(model = "fitted") %>% group_by(state) %>% 
    mutate(rel_prob = prob/max(prob))
) %>%  
  mutate(model = model %>% factor(., levels = c("specified","fitted"))) %>% 
  ungroup()

ggplot(b, aes(x = var1, y = var2, fill = rel_prob))+
  geom_tile() +
  coord_fixed() +
  facet_grid(model ~ state, labeller = "label_both") +
  scale_fill_gradient(low = "white", high = "midnightblue") +
  #scale_fill_gradient(low = "black", high = "mediumspringgreen") +
  theme(axis.text.x = element_text(angle = 90, hjust = 1),
        strip.text = element_text(face = 2)) +
  guides(fill = FALSE)

```
